app.js Fixing a logic error in the POST method handling causing the server to crash if a client sent an empty payload. Also fixing up some documentation.
item.js
Changing 'name' member to 'item' to preserve symmetry across the same
object definition in the Android component.

itemDataSource.js
Implementing an 'arrayObjectIndexOf' method that extends the functionality
of the array 'indexOf()' method, but to search over an array of objects
for a particular value of a particular member.
Modifying 'getItemsSince' method to filter out updated Items that were
received from the client. In this way, a client will not receive the
same Items it sent to the server to be updated.