Finishing up the final portion of version 2 of the API.
An item object is now of the form:
    {
        'id': <object id string>,
        'item': <item name string>,
        'isMarked': <isMarked boolean>,
        'isDeleted': <isDeleted boolean>,
        'timestamp': <Unix timestamp>,
        'version': <version integer>
    }

The most significant change is with the 'id' member: it is now a string
instead of an integer to allow for clients to use guaranteed-unique
item object ID's.
The second modification to the Item object is the addition of the
'version' member. When Items are POSTed to the server, existing item
objects will be compared by the version number instead of the item's
timestamp.