[app.js] Updating the calculation of a response's content-length to be the byte length rather than the character length. This fixes a bug that caused JSON responses to be missing the final character when character length < byte length.
Existing Items are now updated if the POST'ed Item version number is
greater than that of the existing Item. Item timestamps are now only
considered if two Items have equal version numbers. This change is to
reflect the updates to the Groceree API v2.0

[itemDataSource.js]
When updating an Item, the Item's timestamp is set to the server's
current time rather than the timestamp of the POST'ed Item. The server
is now authoritative with respect to time.